Added projectId property to GoogleDocsClient for better configuration management
Applied changes from BB to 1 file

Files modified:
- api/src/dataSources/googledocs/googledocsClient.ts

Added project ID parameter to GoogleDocsClient constructor for configuration access

Applied changes from BB to 1 file

Files modified:
- api/src/dataSources/googledocs/googledocsClient.ts

Added projectId to GoogleDocs client configuration

Applied changes from BB to 1 file

Files modified:
- api/src/dataSources/googledocs/googledocsClient.ts

Added optional projectId parameter to GoogleDocsClient.fromAuthConfig method

Applied changes from BB to 1 file

Files modified:
- api/src/dataSources/googledocs/googledocsClient.ts

Added projectId validation and parameter to GoogleDocsClient constructor for improved authentication and identification

Applied changes from BB to 1 file

Files modified:
- api/src/dataSources/googledocs/googledocsClient.ts

Added config manager import to Google Docs client

Applied changes from BB to 1 file

Files modified:
- api/src/dataSources/googledocs/googledocsClient.ts

Updated token refresh mechanism to use BUI token endpoint with configurable configuration

Applied changes from BB to 1 file

Files modified:
- api/src/dataSources/googledocs/googledocsClient.ts

Added support for refresh token and operation in Google OAuth token route

Applied changes from BB to 1 file

Files modified:
- bui/src/routes/api/v1/oauth/google/token.ts

Added support for token refresh operation with separate parameter validation

Applied changes from BB to 1 file

Files modified:
- bui/src/routes/api/v1/oauth/google/token.ts

Added support for token refresh and improved OAuth token exchange logic

Applied changes from BB to 1 file

Files modified:
- bui/src/routes/api/v1/oauth/google/token.ts

Improved error handling for Google OAuth token exchange and refresh operations

Applied changes from BB to 1 file

Files modified:
- bui/src/routes/api/v1/oauth/google/token.ts

Fixed OAuth token exchange logging to use dynamic operation description

Applied changes from BB to 1 file

Files modified:
- bui/src/routes/api/v1/oauth/google/token.ts

Updated error logging and message for general OAuth token operation error handling

Applied changes from BB to 1 file

Files modified:
- bui/src/routes/api/v1/oauth/google/token.ts

Updated OAuth token route documentation to include token refresh flow and clarify PKCE usage

Applied changes from BB to 1 file

Files modified:
- bui/src/routes/api/v1/oauth/google/token.ts

Added token refresh functionality to Google OAuth flow

Applied changes from BB to 1 file

Files modified:
- bui/src/components/GoogleOAuthFlow.tsx

more portable text; google oauth

Refactored Google OAuth callback handler to support GET method and improved error handling

Applied changes from BB to 1 file

Files modified:
- bui/src/routes/api/v1/oauth/google/token.ts
