+++ api/src/dataSources/googledocs/googledocsClient.ts
@@ -2,7 +2,15 @@
  * GoogleDocsClient for interacting with the Google Docs and Drive APIs.
  */
 import { logger } from 'shared/logger.ts';
+import { getConfigManager } from 'shared/config/configManager.ts';
 import type { AuthConfig } from 'api/dataSources/interfaces/authentication.ts';
+import type {
+	GoogleDocsBatchUpdateRequest,
+	GoogleDocsBatchUpdateResponse,
+	GoogleDocument,
+	GoogleDriveFile,
+	GoogleDriveFilesList,
+} from 'api/dataSources/googledocs.types.ts';
 
 // Google APIs library types
 export interface GoogleAuth {
@@ -11,641 +19,6 @@ export interface GoogleAuth {
 	refreshAccessToken(): Promise<{ credentials: { access_token: string; refresh_token?: string; expiry_date?: number } }>;
 }
 
-export interface GoogleDocs {
-	documents: {
-		get(params: { documentId: string }): Promise<{ data: GoogleDocument }>;
-		create(params: { requestBody: { title: string } }): Promise<{ data: GoogleDocument }>;
-		batchUpdate(params: { 
-			documentId: string; 
-			requestBody: { requests: GoogleDocsBatchUpdateRequest[] }
-		}): Promise<{ data: GoogleDocsBatchUpdateResponse }>;
-	};
-}
-
-export interface GoogleDrive {
-	files: {
-		list(params?: {
-			q?: string;
-			pageSize?: number;
-			pageToken?: string;
-			fields?: string;
-		}): Promise<{ data: GoogleDriveFilesList }>;
-		get(params: {
-			fileId: string;
-			fields?: string;
-		}): Promise<{ data: GoogleDriveFile }>;
-	};
-}
-
-// Google Docs API types
-export interface GoogleDocument {
-	documentId: string;
-	title: string;
-	body: GoogleDocumentBody;
-	headers?: Record<string, GoogleDocumentHeader>;
-	footers?: Record<string, GoogleDocumentFooter>;
-	documentStyle?: GoogleDocumentStyle;
-	namedStyles?: GoogleNamedStyles;
-	lists?: Record<string, GoogleList>;
-	revisionId?: string;
-	suggestionsViewMode?: string;
-}
-
-export interface GoogleDocumentBody {
-	content: GoogleStructuralElement[];
-}
-
-export interface GoogleStructuralElement {
-	startIndex: number;
-	endIndex: number;
-	paragraph?: GoogleParagraph;
-	table?: GoogleTable;
-	sectionBreak?: GoogleSectionBreak;
-	tableOfContents?: GoogleTableOfContents;
-}
-
-export interface GoogleParagraph {
-	elements: GoogleParagraphElement[];
-	paragraphStyle?: GoogleParagraphStyle;
-	positionedObjectIds?: string[];
-}
-
-export interface GoogleParagraphElement {
-	startIndex: number;
-	endIndex: number;
-	textRun?: GoogleTextRun;
-	inlineObjectElement?: GoogleInlineObjectElement;
-	pageBreak?: GooglePageBreak;
-	columnBreak?: GoogleColumnBreak;
-	footnoteReference?: GoogleFootnoteReference;
-	horizontalRule?: GoogleHorizontalRule;
-	equation?: GoogleEquation;
-}
-
-export interface GoogleTextRun {
-	content: string;
-	textStyle?: GoogleTextStyle;
-}
-
-export interface GoogleTextStyle {
-	bold?: boolean;
-	italic?: boolean;
-	underline?: boolean;
-	strikethrough?: boolean;
-	smallCaps?: boolean;
-	backgroundColor?: GoogleOptionalColor;
-	foregroundColor?: GoogleOptionalColor;
-	fontSize?: GoogleDimension;
-	weightedFontFamily?: GoogleWeightedFontFamily;
-	baselineOffset?: string;
-	link?: GoogleLink;
-}
-
-export interface GoogleOptionalColor {
-	color?: GoogleColor;
-}
-
-export interface GoogleColor {
-	rgbColor?: GoogleRgbColor;
-}
-
-export interface GoogleRgbColor {
-	red?: number;
-	green?: number;
-	blue?: number;
-}
-
-export interface GoogleDimension {
-	magnitude?: number;
-	unit?: string;
-}
-
-export interface GoogleWeightedFontFamily {
-	fontFamily?: string;
-	weight?: number;
-}
-
-export interface GoogleLink {
-	url?: string;
-	bookmarkId?: string;
-	headingId?: string;
-}
-
-export interface GoogleParagraphStyle {
-	namedStyleType?: string;
-	alignment?: string;
-	lineSpacing?: number;
-	direction?: string;
-	spacingMode?: string;
-	spaceAbove?: GoogleDimension;
-	spaceBelow?: GoogleDimension;
-	borderBetween?: GoogleParagraphBorder;
-	borderTop?: GoogleParagraphBorder;
-	borderBottom?: GoogleParagraphBorder;
-	borderLeft?: GoogleParagraphBorder;
-	borderRight?: GoogleParagraphBorder;
-	indentFirstLine?: GoogleDimension;
-	indentStart?: GoogleDimension;
-	indentEnd?: GoogleDimension;
-	tabStops?: GoogleTabStop[];
-	keepLinesTogether?: boolean;
-	keepWithNext?: boolean;
-	avoidWidowAndOrphan?: boolean;
-	shading?: GoogleShading;
-	headingId?: string;
-}
-
-export interface GoogleParagraphBorder {
-	color?: GoogleOptionalColor;
-	width?: GoogleDimension;
-	padding?: GoogleDimension;
-	dashStyle?: string;
-}
-
-export interface GoogleTabStop {
-	offset?: GoogleDimension;
-	alignment?: string;
-}
-
-export interface GoogleShading {
-	backgroundColor?: GoogleOptionalColor;
-}
-
-export interface GoogleTable {
-	rows?: number;
-	columns?: number;
-	tableRows?: GoogleTableRow[];
-	tableStyle?: GoogleTableStyle;
-}
-
-export interface GoogleTableRow {
-	startIndex?: number;
-	endIndex?: number;
-	tableCells?: GoogleTableCell[];
-	tableRowStyle?: GoogleTableRowStyle;
-}
-
-export interface GoogleTableCell {
-	startIndex?: number;
-	endIndex?: number;
-	content?: GoogleStructuralElement[];
-	tableCellStyle?: GoogleTableCellStyle;
-}
-
-export interface GoogleTableStyle {
-	tableColumnProperties?: GoogleTableColumnProperties[];
-}
-
-export interface GoogleTableColumnProperties {
-	width?: GoogleDimension;
-	widthType?: string;
-}
-
-export interface GoogleTableRowStyle {
-	minRowHeight?: GoogleDimension;
-}
-
-export interface GoogleTableCellStyle {
-	rowSpan?: number;
-	columnSpan?: number;
-	backgroundColor?: GoogleOptionalColor;
-	borderTop?: GoogleTableCellBorder;
-	borderBottom?: GoogleTableCellBorder;
-	borderLeft?: GoogleTableCellBorder;
-	borderRight?: GoogleTableCellBorder;
-	paddingTop?: GoogleDimension;
-	paddingBottom?: GoogleDimension;
-	paddingLeft?: GoogleDimension;
-	paddingRight?: GoogleDimension;
-	contentAlignment?: string;
-}
-
-export interface GoogleTableCellBorder {
-	color?: GoogleOptionalColor;
-	width?: GoogleDimension;
-	dashStyle?: string;
-}
-
-export interface GoogleSectionBreak {
-	sectionStyle?: GoogleSectionStyle;
-}
-
-export interface GoogleSectionStyle {
-	columnSeparatorStyle?: string;
-	contentDirection?: string;
-	sectionType?: string;
-	defaultHeaderId?: string;
-	defaultFooterId?: string;
-	evenPageHeaderId?: string;
-	evenPageFooterId?: string;
-	firstPageHeaderId?: string;
-	firstPageFooterId?: string;
-	marginTop?: GoogleDimension;
-	marginBottom?: GoogleDimension;
-	marginRight?: GoogleDimension;
-	marginLeft?: GoogleDimension;
-	pageNumberStart?: number;
-	columnProperties?: GoogleSectionColumnProperties[];
-}
-
-export interface GoogleSectionColumnProperties {
-	width?: GoogleDimension;
-	paddingEnd?: GoogleDimension;
-}
-
-export interface GoogleTableOfContents {
-	content?: GoogleStructuralElement[];
-}
-
-export interface GoogleDocumentHeader {
-	content?: GoogleStructuralElement[];
-	headerId?: string;
-}
-
-export interface GoogleDocumentFooter {
-	content?: GoogleStructuralElement[];
-	footerId?: string;
-}
-
-export interface GoogleDocumentStyle {
-	background?: GoogleBackground;
-	pageNumberStart?: number;
-	marginTop?: GoogleDimension;
-	marginBottom?: GoogleDimension;
-	marginRight?: GoogleDimension;
-	marginLeft?: GoogleDimension;
-	pageSize?: GoogleSize;
-	marginHeader?: GoogleDimension;
-	marginFooter?: GoogleDimension;
-	useFirstPageHeaderFooter?: boolean;
-	useEvenPageHeaderFooter?: boolean;
-	flipPageOrientation?: boolean;
-	defaultHeaderId?: string;
-	defaultFooterId?: string;
-	evenPageHeaderId?: string;
-	evenPageFooterId?: string;
-	firstPageHeaderId?: string;
-	firstPageFooterId?: string;
-}
-
-export interface GoogleBackground {
-	color?: GoogleOptionalColor;
-}
-
-export interface GoogleSize {
-	width?: GoogleDimension;
-	height?: GoogleDimension;
-}
-
-export interface GoogleNamedStyles {
-	styles?: GoogleNamedStyle[];
-}
-
-export interface GoogleNamedStyle {
-	namedStyleType?: string;
-	textStyle?: GoogleTextStyle;
-	paragraphStyle?: GoogleParagraphStyle;
-}
-
-export interface GoogleList {
-	listProperties?: GoogleListProperties;
-}
-
-export interface GoogleListProperties {
-	nestingLevels?: GoogleNestingLevel[];
-}
-
-export interface GoogleNestingLevel {
-	bulletAlignment?: string;
-	glyphFormat?: string;
-	glyphSymbol?: string;
-	glyphType?: string;
-	indentFirstLine?: GoogleDimension;
-	indentStart?: GoogleDimension;
-	textStyle?: GoogleTextStyle;
-	startNumber?: number;
-}
-
-// Additional element types
-export interface GoogleInlineObjectElement {
-	inlineObjectId?: string;
-	textStyle?: GoogleTextStyle;
-}
-
-export interface GooglePageBreak {
-	textStyle?: GoogleTextStyle;
-}
-
-export interface GoogleColumnBreak {
-	textStyle?: GoogleTextStyle;
-}
-
-export interface GoogleFootnoteReference {
-	footnoteId?: string;
-	footnoteNumber?: string;
-	textStyle?: GoogleTextStyle;
-}
-
-export interface GoogleHorizontalRule {
-	textStyle?: GoogleTextStyle;
-}
-
-export interface GoogleEquation {
-	textStyle?: GoogleTextStyle;
-}
-
-// Batch update types
-export interface GoogleDocsBatchUpdateRequest {
-	insertText?: {
-		location: { index: number };
-		text: string;
-	};
-	deleteContentRange?: {
-		range: { startIndex: number; endIndex: number };
-	};
-	updateTextStyle?: {
-		range: { startIndex: number; endIndex: number };
-		textStyle: GoogleTextStyle;
-		fields: string;
-	};
-	updateParagraphStyle?: {
-		range: { startIndex: number; endIndex: number };
-		paragraphStyle: GoogleParagraphStyle;
-		fields: string;
-	};
-	insertPageBreak?: {
-		location: { index: number };
-	};
-	insertSectionBreak?: {
-		location: { index: number };
-		sectionType: string;
-	};
-	insertTable?: {
-		location: { index: number };
-		rows: number;
-		columns: number;
-	};
-	insertTableRow?: {
-		tableCellLocation: {
-			tableStartLocation: { index: number };
-			rowIndex: number;
-			columnIndex: number;
-		};
-		insertBelow: boolean;
-	};
-	insertTableColumn?: {
-		tableCellLocation: {
-			tableStartLocation: { index: number };
-			rowIndex: number;
-			columnIndex: number;
-		};
-		insertRight: boolean;
-	};
-	deleteTableRow?: {
-		tableCellLocation: {
-			tableStartLocation: { index: number };
-			rowIndex: number;
-			columnIndex: number;
-		};
-	};
-	deleteTableColumn?: {
-		tableCellLocation: {
-			tableStartLocation: { index: number };
-			rowIndex: number;
-			columnIndex: number;
-		};
-	};
-	replaceAllText?: {
-		containsText: {
-			text: string;
-			matchCase: boolean;
-		};
-		replaceText: string;
-	};
-	createParagraphBullets?: {
-		range: { startIndex: number; endIndex: number };
-		bulletPreset: string;
-	};
-	deleteParagraphBullets?: {
-		range: { startIndex: number; endIndex: number };
-	};
-	createNamedRange?: {
-		name: string;
-		range: { startIndex: number; endIndex: number };
-	};
-	deleteNamedRange?: {
-		name?: string;
-		namedRangeId?: string;
-	};
-}
-
-export interface GoogleDocsBatchUpdateResponse {
-	documentId: string;
-	replies: GoogleDocsBatchUpdateReply[];
-}
-
-export interface GoogleDocsBatchUpdateReply {
-	insertText?: {};
-	deleteContentRange?: {};
-	updateTextStyle?: {};
-	updateParagraphStyle?: {};
-	insertPageBreak?: {};
-	insertSectionBreak?: {};
-	insertTable?: {};
-	insertTableRow?: {};
-	insertTableColumn?: {};
-	deleteTableRow?: {};
-	deleteTableColumn?: {};
-	replaceAllText?: {
-		occurrencesChanged: number;
-	};
-	createParagraphBullets?: {};
-	deleteParagraphBullets?: {};
-	createNamedRange?: {
-		namedRangeId: string;
-	};
-	deleteNamedRange?: {};
-}
-
-// Google Drive API types
-export interface GoogleDriveFilesList {
-	kind: string;
-	nextPageToken?: string;
-	incompleteSearch?: boolean;
-	files: GoogleDriveFile[];
-}
-
-export interface GoogleDriveFile {
-	kind: string;
-	id: string;
-	name: string;
-	mimeType: string;
-	description?: string;
-	starred?: boolean;
-	trashed?: boolean;
-	explicitlyTrashed?: boolean;
-	parents?: string[];
-	properties?: Record<string, string>;
-	appProperties?: Record<string, string>;
-	spaces?: string[];
-	version?: string;
-	webContentLink?: string;
-	webViewLink?: string;
-	iconLink?: string;
-	hasThumbnail?: boolean;
-	thumbnailLink?: string;
-	thumbnailVersion?: string;
-	viewedByMe?: boolean;
-	viewedByMeTime?: string;
-	createdTime?: string;
-	modifiedTime?: string;
-	modifiedByMeTime?: string;
-	modifiedByMe?: boolean;
-	sharedWithMeTime?: string;
-	sharingUser?: GoogleDriveUser;
-	owners?: GoogleDriveUser[];
-	teamDriveId?: string;
-	driveId?: string;
-	lastModifyingUser?: GoogleDriveUser;
-	shared?: boolean;
-	ownedByMe?: boolean;
-	capabilities?: GoogleDriveCapabilities;
-	viewersCanCopyContent?: boolean;
-	copyRequiresWriterPermission?: boolean;
-	writersCanShare?: boolean;
-	permissions?: GoogleDrivePermission[];
-	permissionIds?: string[];
-	hasAugmentedPermissions?: boolean;
-	folderColorRgb?: string;
-	originalFilename?: string;
-	fullFileExtension?: string;
-	fileExtension?: string;
-	md5Checksum?: string;
-	sha1Checksum?: string;
-	sha256Checksum?: string;
-	size?: string;
-	quotaBytesUsed?: string;
-	headRevisionId?: string;
-	contentHints?: {
-		thumbnail?: {
-			image?: string;
-			mimeType?: string;
-		};
-		indexableText?: string;
-	};
-	imageMediaMetadata?: {
-		width?: number;
-		height?: number;
-		rotation?: number;
-		location?: {
-			latitude?: number;
-			longitude?: number;
-			altitude?: number;
-		};
-		time?: string;
-		cameraMake?: string;
-		cameraModel?: string;
-		exposureTime?: number;
-		aperture?: number;
-		flashUsed?: boolean;
-		focalLength?: number;
-		isoSpeed?: number;
-		meteringMode?: string;
-		sensor?: string;
-		exposureMode?: string;
-		colorSpace?: string;
-		whiteBalance?: string;
-		exposureBias?: number;
-		maxApertureValue?: number;
-		subjectDistance?: number;
-		lens?: string;
-	};
-	videoMediaMetadata?: {
-		width?: number;
-		height?: number;
-		durationMillis?: string;
-	};
-	isAppAuthorized?: boolean;
-	exportLinks?: Record<string, string>;
-}
-
-export interface GoogleDriveUser {
-	kind: string;
-	displayName: string;
-	photoLink?: string;
-	me?: boolean;
-	permissionId: string;
-	emailAddress?: string;
-}
-
-export interface GoogleDriveCapabilities {
-	canAcceptOwnership?: boolean;
-	canAddChildren?: boolean;
-	canAddFolderFromAnotherDrive?: boolean;
-	canAddMyDriveParent?: boolean;
-	canChangeCopyRequiresWriterPermission?: boolean;
-	canChangeSecurityUpdateEnabled?: boolean;
-	canChangeViewersCanCopyContent?: boolean;
-	canComment?: boolean;
-	canCopy?: boolean;
-	canCreate?: boolean;
-	canDelete?: boolean;
-	canDeleteChildren?: boolean;
-	canDownload?: boolean;
-	canEdit?: boolean;
-	canListChildren?: boolean;
-	canModifyContent?: boolean;
-	canModifyContentRestriction?: boolean;
-	canMoveChildrenOutOfTeamDrive?: boolean;
-	canMoveChildrenOutOfDrive?: boolean;
-	canMoveChildrenWithinTeamDrive?: boolean;
-	canMoveChildrenWithinDrive?: boolean;
-	canMoveItemIntoTeamDrive?: boolean;
-	canMoveItemOutOfTeamDrive?: boolean;
-	canMoveItemOutOfDrive?: boolean;
-	canMoveItemWithinTeamDrive?: boolean;
-	canMoveItemWithinDrive?: boolean;
-	canMoveTeamDriveItem?: boolean;
-	canReadLabels?: boolean;
-	canReadRevisions?: boolean;
-	canReadTeamDrive?: boolean;
-	canReadDrive?: boolean;
-	canRemoveChildren?: boolean;
-	canRemoveMyDriveParent?: boolean;
-	canRename?: boolean;
-	canShare?: boolean;
-	canTrash?: boolean;
-	canTrashChildren?: boolean;
-	canUntrash?: boolean;
-}
-
-export interface GoogleDrivePermission {
-	kind: string;
-	id: string;
-	type: string;
-	emailAddress?: string;
-	domain?: string;
-	role: string;
-	view?: string;
-	allowFileDiscovery?: boolean;
-	displayName?: string;
-	photoLink?: string;
-	expirationTime?: string;
-	teamDrivePermissionDetails?: Array<{
-		teamDrivePermissionType: string;
-		role: string;
-		inherited: boolean;
-		inheritedFrom: string;
-	}>;
-	permissionDetails?: Array<{
-		permissionType: string;
-		role: string;
-		inherited: boolean;
-		inheritedFrom: string;
-	}>;
-	deleted?: boolean;
-	pendingOwner?: boolean;
-}
 
 /**
  * Client for interacting with the Google Docs and Drive APIs
@@ -656,64 +29,62 @@ export class GoogleDocsClient {
 	private clientId: string;
 	private clientSecret: string;
 	private expiresAt?: number;
+	private projectId: string;
 	private readonly apiBaseUrl = 'https://www.googleapis.com';
 	private readonly authUrl = 'https://oauth2.googleapis.com/token';
-	private readonly requiredScopes = [
-		'https://www.googleapis.com/auth/documents',
-		'https://www.googleapis.com/auth/drive.readonly',
-		'https://www.googleapis.com/auth/drive.file'
-	];
 
 	/**
 	 * Create a new GoogleDocsClient
+	 * @param clientId OAuth2 client ID
 	 * @param accessToken OAuth2 access token
 	 * @param refreshToken OAuth2 refresh token (optional)
-	 * @param clientId OAuth2 client ID
-	 * @param clientSecret OAuth2 client secret
 	 * @param expiresAt Token expiration timestamp
+	 * @param projectId Project ID for configuration access
 	 */
 	constructor(
+		clientId: string,
 		accessToken: string,
 		refreshToken: string | undefined,
-		clientId: string,
-		clientSecret: string,
-		expiresAt?: number
+		expiresAt: number | undefined,
+		projectId: string,
 	) {
 		this.accessToken = accessToken;
 		this.refreshToken = refreshToken;
-		this.clientId = clientId;
-		this.clientSecret = clientSecret;
 		this.expiresAt = expiresAt;
+		this.clientId = clientId;
+		this.clientSecret = ''; //clientSecret;
+		this.projectId = projectId;
 	}
 
 	/**
 	 * Create a GoogleDocsClient from an AuthConfig
 	 * @param auth Authentication configuration
+	 * @param projectId Project ID for configuration access
 	 * @returns A new GoogleDocsClient instance or null if auth is invalid
 	 */
-	static fromAuthConfig(auth?: AuthConfig): GoogleDocsClient | null {
+	static fromAuthConfig(auth?: AuthConfig, projectId?: string): GoogleDocsClient | null {
 		if (!auth || auth.method !== 'oauth2') {
 			logger.warn('GoogleDocsClient: Invalid auth config, must use oauth2 method');
 			return null;
 		}
 
-		const accessToken = auth.credentials?.accessToken as string;
-		const clientId = auth.credentials?.clientId as string;
-		const clientSecret = auth.credentials?.clientSecret as string;
-		const refreshToken = auth.tokenData?.refreshToken;
-		const expiresAt = auth.tokenData?.expiresAt;
+		const clientId = ''; //auth.oauth2?.clientId as string;
+		//const clientSecret = ''; //auth.oauth2?.clientSecret as string;
+		const accessToken = auth.oauth2?.accessToken as string;
+		const refreshToken = auth.oauth2?.refreshToken;
+		const expiresAt = auth.oauth2?.expiresAt;
 
-		if (!accessToken || !clientId || !clientSecret) {
-			logger.warn('GoogleDocsClient: Missing required OAuth2 credentials');
+		if (!accessToken || !clientId || !projectId) {
+			logger.warn('GoogleDocsClient: Missing required OAuth2 credentials or projectId');
 			return null;
 		}
 
 		return new GoogleDocsClient(
+			clientId,
 			accessToken,
 			refreshToken,
-			clientId,
-			clientSecret,
-			expiresAt
+			expiresAt,
+			projectId,
 		);
 	}
 
@@ -731,7 +102,7 @@ export class GoogleDocsClient {
 	}
 
 	/**
-	 * Refresh the access token using the refresh token
+	 * Refresh the access token using the refresh token via BUI endpoint
 	 * @returns True if refresh was successful
 	 */
 	private async refreshAccessToken(): Promise<boolean> {
@@ -741,16 +112,25 @@ export class GoogleDocsClient {
 		}
 
 		try {
-			const response = await fetch(this.authUrl, {
+			// Get token endpoint from configuration
+			const configManager = await getConfigManager();
+			const globalConfig = await configManager.getGlobalConfig();
+			const projectConfig = await configManager.getProjectConfig(this.projectId);
+			const tokenEndpoint = projectConfig.api.tokenEndpoint || globalConfig.api.tokenEndpoint;
+			
+			if (!tokenEndpoint) {
+				logger.error('GoogleDocsClient: No token endpoint configured');
+				return false;
+			}
+			
+			const response = await fetch(tokenEndpoint, {
 				method: 'POST',
 				headers: {
-					'Content-Type': 'application/x-www-form-urlencoded',
+					'Content-Type': 'application/json',
 				},
-				body: new URLSearchParams({
-					grant_type: 'refresh_token',
-					refresh_token: this.refreshToken,
-					client_id: this.clientId,
-					client_secret: this.clientSecret,
+				body: JSON.stringify({
+					refreshToken: this.refreshToken,
+					operation: 'refresh',
 				}),
 			});
 
@@ -761,16 +141,16 @@ export class GoogleDocsClient {
 			}
 
 			const data = await response.json();
-			this.accessToken = data.access_token;
+			this.accessToken = data.accessToken;
 			
 			// Update expiry time if provided
-			if (data.expires_in) {
-				this.expiresAt = Date.now() + (data.expires_in * 1000);
+			if (data.expiresIn) {
+				this.expiresAt = Date.now() + (data.expiresIn * 1000);
 			}
 
 			// Update refresh token if provided
-			if (data.refresh_token) {
-				this.refreshToken = data.refresh_token;
+			if (data.refreshToken) {
+				this.refreshToken = data.refreshToken;
 			}
 
 			logger.info('GoogleDocsClient: Successfully refreshed access token');
@@ -929,7 +309,8 @@ export class GoogleDocsClient {
 		let q = "mimeType='application/vnd.google-apps.document'";
 		
 		if (query) {
-			q += ` and name contains '${query.replace(/'/g, "\\'")}'`;
+			const escapedQuery = query.replace(/'/g, "\\'");
+			q += ` and name contains '${escapedQuery}'`;
 		}
 		
 		if (folderId) {
