+++ api/src/dataSources/interfaces/authentication.ts
@@ -3,6 +3,8 @@
  * Authentication is a cross-cutting concern that affects multiple components in the data source system.
  */
 
+import type { DataSourceAuth } from 'shared/types/dataSource.ts';
+
 /**
  * Supported authentication methods for data sources
  */
@@ -17,43 +19,54 @@ export type AuthMethod =
 /**
  * Authentication configuration for data sources
  */
-export interface AuthConfig {
-	/**
-	 * The authentication method to use
-	 */
-	method: AuthMethod;
-
-	/**
-	 * API key for 'apiKey' method
-	 */
-	apiKey?: string;
-
-	/**
-	 * Credentials for various auth methods
-	 * Stored as key-value pairs appropriate to the method
-	 */
-	credentials?: Record<string, unknown>;
-
-	/**
-	 * OAuth token data when using 'oauth2' method
-	 */
-	tokenData?: {
-		/**
-		 * Timestamp when the token expires (in milliseconds since epoch)
-		 */
-		expiresAt: number;
-
-		/**
-		 * Refresh token for obtaining a new access token
-		 */
-		refreshToken?: string;
-
-		/**
-		 * Scope of access granted by the token
-		 */
-		scope?: string;
-	};
-}
+// Important - there is also DataSourceAuth type in shared/types/dataSource.ts - which is the canonical definition??
+export type AuthConfig = DataSourceAuth;
+// export interface AuthConfig {
+// 	/**
+// 	 * The authentication method to use
+// 	 */
+// 	method: AuthMethod;
+// 
+// 	/**
+// 	 * API key for 'apiKey' method
+// 	 */
+// 	apiKey?: string;
+// 
+// 	/**
+// 	 * Credentials for various auth methods
+// 	 */
+// 
+// 	// References to future secure credential storage
+// 	// These would be keys to look up in a secure storage mechanism
+// 	credentialRefs?: string[];
+// 
+// 	basic?: DataSourceAuthBasic;
+// 	bearer?: DataSourceAuthBearer;
+// 
+// 	/**
+// 	 * OAuth token data when using 'oauth2' method
+// 	 */
+// 	oauth2?: {
+// 		//clientId: string;
+// 		//clientSecret: string;
+// 		//tokenData?: {
+// 			accessToken: string;
+// 			/**
+// 			 * Refresh token for obtaining a new access token
+// 			 */
+// 			refreshToken?: string;
+// 			/**
+// 			 * Timestamp when the token expires (in milliseconds since epoch)
+// 			 */
+// 			expiresAt: number;
+// 			/**
+// 			 * Scope of access granted by the token
+// 			 */
+// 			scopes?: string;
+// 			tokenType?: string;
+// 		//};
+// 	};
+// }
 
 /**
  * OAuth-specific configuration
