+++ import_map.json
@@ -148,6 +148,11 @@
 		"api/dataSources/notionClient.ts": "./api/src/dataSources/notion/notionClient.ts",
 		"api/dataSources/notionAccessor.ts": "./api/src/dataSources/notion/notionAccessor.ts",
 		"api/dataSources/notion/portableTextConverter.ts": "./api/src/dataSources/notion/portableTextConverter.ts",
+		"api/dataSources/googledocsProvider.ts": "./api/src/dataSources/googledocs/googledocsProvider.ts",
+		"api/dataSources/googledocsClient.ts": "./api/src/dataSources/googledocs/googledocsClient.ts",
+		"api/dataSources/googledocsAccessor.ts": "./api/src/dataSources/googledocs/googledocsAccessor.ts",
+		"api/dataSources/googledocs/portableTextConverter.ts": "./api/src/dataSources/googledocs/portableTextConverter.ts",
+		"api/dataSources/googledocs.types.ts": "./api/src/dataSources/googledocs/googledocstypes.ts",
 		"api/dataSources/genericMCPProvider.ts": "./api/src/dataSources/mcp/genericMCPProvider.ts",
 		"api/resources/resourceManager.ts": "./api/src/resources/resourceManager.ts",
 		"api/logEntries/types.ts": "./api/src/logEntries/types.ts",
