+++ src/shared/types/dataSource.types.ts
@@ -5,7 +5,8 @@ export type DataSourceProviderType =
 	| 'filesystem' // Local file system
 	| 'database' // Database connections
 	| 'notion' // Notion workspaces
-	| 'gdrive' // Google Drive
+	| 'googledocs' // Google Drive
+	//| 'supabase' // Supabase projects
 	| 'mcp' // Model Context Protocol servers
 	| string; // Future extensions and MCP types
 
@@ -26,49 +27,70 @@ export interface DataSourceMetadata {
 	totalResources: number;
 	resourceTypes: Record<string, number>;
 	lastScanned: string;
-	filesystem?: {
-		totalDirectories: number;
-		totalFiles: number;
-		largestFileSize?: number;
-		deepestPathDepth?: number;
-		fileExtensions?: Record<string, number>;
-		oldestFileDate?: string;
-		newestFileDate?: string;
-		capabilities?: {
-			canRead: boolean; // We can scan, so we can read
-			canWrite: boolean; // Will test below
-			canDelete: boolean; // Will test below
-			canMove: boolean; // Will test below
-			hasRestrictedAreas: boolean; // Will detect below
-		};
-		contentVisibility?: {
-			includesHiddenFiles: boolean; // Based on our exclude patterns
-			includesDotDirectories: boolean; // We skip .git, .bb, etc.
-			followsSymlinks: boolean; // walk options set includeSymlinks: false
-			brokenSymlinkCount: number;
-			filteredByGitignore: boolean; // We use getExcludeOptions
-			filteredByBBIgnore: boolean;
-		};
-		practicalLimits?: {
-			maxFileSize: number; // 10MB reasonable limit for text processing
-			recommendedPageSize: number; // Good balance for filesystem
-			hasVeryLargeFiles: boolean; // Will detect below
-		};
-		contentAnalysis?: {
-			textFileCount: number;
-			binaryFileCount: number;
-			likelyEncodingIssues: number;
-			emptyFileCount: number;
-		};
+	filesystem?: DataSourceMetadataFilesystem;
+	notion?: DataSourceMetadataNotion;
+	googledocs?: DataSourceMetadataGoogleDocs;
+}
+
+// Filesystem-specific metadata
+export interface DataSourceMetadataFilesystem {
+	totalDirectories: number;
+	totalFiles: number;
+	largestFileSize?: number;
+	deepestPathDepth?: number;
+	fileExtensions?: Record<string, number>; // e.g., { '.ts': 50, '.md': 10, '.json': 5 }
+	oldestFileDate?: string; // ISO 8601 string
+	newestFileDate?: string; // ISO 8601 string
+	// LLM-critical operational metadata
+	capabilities?: {
+		canRead: boolean; // We can scan, so we can read
+		canWrite: boolean; // Will test below
+		canDelete: boolean; // Will test below
+		canMove: boolean; // Will test below
+		hasRestrictedAreas: boolean; // Will detect below - Some areas may be inaccessible
 	};
-	notion?: {
-		totalPages: number;
-		totalDatabases: number;
-		pageTypes: Record<string, number>;
-		workspaceInfo?: {
-			name: string;
-			id: string;
-		};
+	contentVisibility?: {
+		includesHiddenFiles: boolean; // Based on our exclude patterns - Files starting with .
+		includesDotDirectories: boolean; // We skip .git, .bb, etc.
+		followsSymlinks: boolean; // walk options set includeSymlinks: false
+		brokenSymlinkCount: number;
+		filteredByGitignore: boolean; // We use getExcludeOptions
+		filteredByBBIgnore: boolean;
+	};
+	practicalLimits?: {
+		maxFileSize: number; // Largest file that can be processed - 10MB reasonable limit for text processing
+		recommendedPageSize: number; // Optimal pageSize for this datasource - Good balance for filesystem
+		hasVeryLargeFiles: boolean; // Files that might timeout/fail to load - Will detect below
+	};
+	contentAnalysis?: {
+		textFileCount: number;
+		binaryFileCount: number;
+		likelyEncodingIssues: number; // Files that might have encoding problems
+		emptyFileCount: number;
+	};
+}
+
+// Notion-specific metadata
+export interface DataSourceMetadataNotion {
+	totalPages: number;
+	totalDatabases: number;
+	pageTypes: Record<string, number>; // e.g., { 'database': 5, 'page': 25 }
+	workspaceInfo?: {
+		name: string;
+		id: string;
+	};
+}
+
+// GoogleDocs-specific metadata
+export interface DataSourceMetadataGoogleDocs {
+	totalDocuments: number;
+	folderId?: string;
+	driveId?: string;
+
+	documentTypes?: Record<string, number>; // e.g., { 'text': 5, 'spreadsheet': 25 }
+	workspaceInfo?: {
+		name: string;
+		id: string;
 	};
 }
 
@@ -83,7 +105,7 @@ export type DataSourceAccessMethod =
 /**
  * Capabilities for data sources
  */
-export type DataSourceCapability = 'read' | 'write' | 'list' | 'search' | 'move' | 'delete'; // Delegated to MCP server
+export type DataSourceCapability = 'read' | 'write' | 'list' | 'search' | 'move' | 'delete' | 'blockRead' | 'blockEdit'; // Delegated to MCP server
 
 /**
  * Available authentication methods
@@ -106,29 +128,27 @@ export interface DataSourceAuthBasic {
 export interface DataSourceAuthBearer {
 	tokenRef: string;
 }
+export interface DataSourceAuthOauth2 {
+	tokenType?: string;
+	accessToken: string;
+	refreshToken?: string;
+	expiresAt?: number;
+	scopes?: string;
+}
+
+// Important - there is also AuthConfig interface in api/dataSources/interfaces/authentication.ts - which is the canonical definition??
 export interface DataSourceAuth {
 	method: DataSourceAuthMethod; // Authentication method
 	apiKey?: string; // For apiKey auth (simple implementation)
 
-	// References to future secure credential storage
-	// These would be keys to look up in a secure storage mechanism
-	credentialRefs?: string[];
-
 	basic?: DataSourceAuthBasic;
 	bearer?: DataSourceAuthBearer;
 
-	// Placeholder fields for future auth methods
-	// These would be implemented later with proper secure storage
-	/*
-	oauth2?: {
-		clientId: string;
-		tokenData?: {
-			expiresAt: number;
-			scope: string;
-			tokenType: string;
-		};
-	};
-	 */
+	oauth2?: DataSourceAuthOauth2;
+
+	// References to future secure credential storage
+	// These would be keys to look up in a secure storage mechanism
+	credentialRefs?: string[];
 }
 
 /**
